src/ingest_data_api_to_s3_pl/requirements.txt
boto3
python-dotenv
polars
snowflake-connector-python
s3fs